+++ cloud-sql/mysql/Startup.cs
@@ -21,17 +21,12 @@ using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
 using MySql.Data.MySqlClient;
-using System;
-using System.Data.Common;
-using System.Data;
-using Polly;
 
 namespace CloudSql
 {
     public class Startup
     {
         public IConfiguration Configuration { get; }
-        IServiceCollection _services;
 
         public Startup(IConfiguration configuration)
         {
@@ -54,157 +49,11 @@ namespace CloudSql
                     options.Version = "Test";
                 });
             }
-            services.AddScoped(typeof(MySqlConnectionStringBuilder),
-                (IServiceProvider) => new MySqlConnect().GetMySqlConnectionString());
-            services.AddScoped<MySqlConnect>();
+            services.AddScoped(sp => new MySqlConnection().GetMySqlConnectionString());
             services.AddMvc(options =>
             {
                 options.Filters.Add(typeof(DbExceptionFilterAttribute));
             });
-            _services = services;
-        }
-
-        public class MySqlConnect
-        {
-            public DbConnection GetMySqlConnection(
-                MySqlConnectionStringBuilder connectionString)
-            {
-                // [START cloud_sql_mysql_dotnet_ado_backoff]
-                var connection = Policy
-                    .HandleResult<DbConnection>(conn => conn.State != ConnectionState.Open)
-                    .WaitAndRetry(new[]
-                    {
-                    TimeSpan.FromSeconds(1),
-                    TimeSpan.FromSeconds(2),
-                    TimeSpan.FromSeconds(5)
-                    }, (result, timeSpan, retryCount, context) =>
-                    {
-                        // Log any warnings here.
-                    })
-                    .Execute(() =>
-                    {
-                        // Return a new connection.
-                        return new MySqlConnection(connectionString.ConnectionString);
-                    });
-                // [END cloud_sql_mysql_dotnet_ado_backoff]
-                return connection;
-            }
-
-            public void InitializeDatabase(MySqlConnectionStringBuilder connectionString)
-            {
-                DbConnection connection;
-                using (connection =
-                    new MySqlConnect().GetMySqlConnection(connectionString))
-                {
-                    connection.Open();
-                    using (var createTableCommand = connection.CreateCommand())
-                    {
-                        createTableCommand.CommandText = @"
-                            CREATE TABLE IF NOT EXISTS
-                            votes(
-                                vote_id SERIAL NOT NULL,
-                                time_cast timestamp NOT NULL,
-                                candidate CHAR(6) NOT NULL,
-                                PRIMARY KEY (vote_id)
-                            )";
-                        createTableCommand.ExecuteNonQuery();
-                    }
-                }
-            }
-
-            void SetDbConfigOptions(MySqlConnectionStringBuilder connectionString)
-            {
-                // [START cloud_sql_mysql_dotnet_ado_limit]
-                // MaximumPoolSize sets maximum number of connections allowed in the pool.
-                connectionString.MaximumPoolSize = 5;
-                // MinimumPoolSize sets the minimum number of connections in the pool.
-                connectionString.MinimumPoolSize = 0;
-                // [END cloud_sql_mysql_dotnet_ado_limit]
-                // [START cloud_sql_mysql_dotnet_ado_timeout]
-                // ConnectionTimeout sets the time to wait (in seconds) while
-                // trying to establish a connection before terminating the attempt.
-                connectionString.ConnectionTimeout = 15;
-                // [END cloud_sql_mysql_dotnet_ado_timeout]
-                // [START cloud_sql_mysql_dotnet_ado_lifetime]
-                // ConnectionLifeTime sets the lifetime of a pooled connection
-                // (in seconds) that a connection lives before it is destroyed
-                // and recreated. Connections that are returned to the pool are
-                // destroyed if it's been more than the number of seconds
-                // specified by ConnectionLifeTime since the connection was
-                // created. The default value is zero (0) which means the
-                // connection always returns to pool.
-                connectionString.ConnectionLifeTime = 1800; // 30 minutes
-                // [END cloud_sql_mysql_dotnet_ado_lifetime]
-            }
-
-            public MySqlConnectionStringBuilder NewMysqlTCPConnectionString()
-            {
-                // [START cloud_sql_mysql_dotnet_ado_connection_tcp]
-                // Equivalent connection string:
-                // "Uid=<DB_USER>;Pwd=<DB_PASS>;Host=<DB_HOST>;Database=<DB_NAME>;"
-                var connectionString = new MySqlConnectionStringBuilder()
-                {
-                    // The Cloud SQL proxy provides encryption between the proxy and instance.
-                    SslMode = MySqlSslMode.None,
-
-                    // Remember - storing secrets in plaintext is potentially unsafe. Consider using
-                    // something like https://cloud.google.com/secret-manager/docs/overview to help keep
-                    // secrets secret.
-                    Server = Environment.GetEnvironmentVariable("DB_HOST"),   // e.g. '127.0.0.1'
-                    // Set Host to 'cloudsql' when deploying to App Engine Flexible environment
-                    UserID = Environment.GetEnvironmentVariable("DB_USER"),   // e.g. 'my-db-user'
-                    Password = Environment.GetEnvironmentVariable("DB_PASS"), // e.g. 'my-db-password'
-                    Database = Environment.GetEnvironmentVariable("DB_NAME"), // e.g. 'my-database'
-                };
-                connectionString.Pooling = true;
-                // Specify additional properties here.
-                // [START_EXCLUDE]
-                SetDbConfigOptions(connectionString);
-                // [END_EXCLUDE]
-                return connectionString;
-                // [END cloud_sql_mysql_dotnet_ado_connection_tcp]
-            }
-
-            public MySqlConnectionStringBuilder NewMysqlUnixSocketConnectionString()
-            {
-                // [START cloud_sql_mysql_dotnet_ado_connection_socket]
-                // Equivalent connection string:
-                // "Server=<dbSocketDir>/<INSTANCE_CONNECTION_NAME>;Uid=<DB_USER>;Pwd=<DB_PASS>;Database=<DB_NAME>;Protocol=unix"
-                String dbSocketDir = Environment.GetEnvironmentVariable("DB_SOCKET_PATH") ?? "/cloudsql";
-                String instanceConnectionName = Environment.GetEnvironmentVariable("INSTANCE_CONNECTION_NAME");
-                var connectionString = new MySqlConnectionStringBuilder()
-                {
-                    // The Cloud SQL proxy provides encryption between the proxy and instance.
-                    SslMode = MySqlSslMode.None,
-                    // Remember - storing secrets in plaintext is potentially unsafe. Consider using
-                    // something like https://cloud.google.com/secret-manager/docs/overview to help keep
-                    // secrets secret.
-                    Server = String.Format("{0}/{1}", dbSocketDir, instanceConnectionName),
-                    UserID = Environment.GetEnvironmentVariable("DB_USER"),   // e.g. 'my-db-user
-                    Password = Environment.GetEnvironmentVariable("DB_PASS"), // e.g. 'my-db-password'
-                    Database = Environment.GetEnvironmentVariable("DB_NAME"), // e.g. 'my-database'
-                    ConnectionProtocol = MySqlConnectionProtocol.UnixSocket
-                };
-                connectionString.Pooling = true;
-                // Specify additional properties here.
-                // [START_EXCLUDE]
-                SetDbConfigOptions(connectionString);
-                // [END_EXCLUDE]
-                return connectionString;
-                // [END cloud_sql_mysql_dotnet_ado_connection_socket]
-            }
-
-            public MySqlConnectionStringBuilder GetMySqlConnectionString()
-            {
-                if (Environment.GetEnvironmentVariable("DB_HOST") != null)
-                {
-                    return NewMysqlTCPConnectionString();
-                }
-                else
-                {
-                    return NewMysqlUnixSocketConnectionString();
-                }
-            }
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
