+++ cloud-sql/postgres/Startup.cs
@@ -21,17 +21,12 @@ using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
 using Npgsql;
-using System;
-using System.Data.Common;
-using System.Data;
-using Polly;
 
 namespace CloudSql
 {
     public class Startup
     {
         public IConfiguration Configuration { get; }
-        IServiceCollection _services;
 
         public Startup(IConfiguration configuration)
         {
@@ -54,152 +49,11 @@ namespace CloudSql
                     options.Version = "Test";
                 });
             }
-            services.AddScoped(typeof(NpgsqlConnectionStringBuilder),
-                (IServiceProvider) => new PostgreSqlConnection().GetPostgreSqlConnectionString());
-            services.AddScoped<PostgreSqlConnection>();
+            services.AddScoped(sp => new NpgsqlConnection().GetPostgreSqlConnectionString());
             services.AddMvc(options =>
             {
                 options.Filters.Add(typeof(DbExceptionFilterAttribute));
             });
-            _services = services;
-        }
-
-        public class PostgreSqlConnection
-        {
-            public DbConnection GetPostgreSqlConnection(
-                NpgsqlConnectionStringBuilder connectionString)
-            {
-                // [START cloud_sql_postgres_dotnet_ado_backoff]
-                var connection = Policy
-                    .HandleResult<DbConnection>(conn => conn.State != ConnectionState.Open)
-                    .WaitAndRetry(new[]
-                    {
-                        TimeSpan.FromSeconds(1),
-                        TimeSpan.FromSeconds(2),
-                        TimeSpan.FromSeconds(5)
-                    }, (result, timeSpan, retryCount, context) =>
-                    {
-                        // Log any warnings here.
-                    })
-                    .Execute(() =>
-                    {
-                        // Return a new connection.
-                        return new NpgsqlConnection(connectionString.ConnectionString);
-                    });
-                // [END cloud_sql_postgres_dotnet_ado_backoff]
-                return connection;
-            }
-
-        public void InitializeDatabase(NpgsqlConnectionStringBuilder connectionString)
-            {
-                DbConnection connection;
-                using (connection =
-                    new PostgreSqlConnection().GetPostgreSqlConnection(connectionString))
-                {
-                    connection.Open();
-                    using (var createTableCommand = connection.CreateCommand())
-                    {
-                        createTableCommand.CommandText = @"
-                            CREATE TABLE IF NOT EXISTS
-                            votes(
-                                vote_id SERIAL NOT NULL,
-                                time_cast timestamp NOT NULL,
-                                candidate VARCHAR(6) NOT NULL,
-                                PRIMARY KEY (vote_id)
-                            )";
-                        createTableCommand.ExecuteNonQuery();
-                    }
-                }
-            }
-
-            void SetDbConfigOptions(NpgsqlConnectionStringBuilder connectionString)
-            {
-                // [START cloud_sql_postgres_dotnet_ado_limit]
-                // MaxPoolSize sets maximum number of connections allowed in the pool.
-                connectionString.MaxPoolSize = 5;
-                // MinPoolSize sets the minimum number of connections in the pool.
-                connectionString.MinPoolSize = 0;
-                // [END cloud_sql_postgres_dotnet_ado_limit]
-                // [START cloud_sql_postgres_dotnet_ado_timeout]
-                // Timeout sets the time to wait (in seconds) while
-                // trying to establish a connection before terminating the attempt.
-                connectionString.Timeout = 15;
-                // [END cloud_sql_postgres_dotnet_ado_timeout]
-                // [START cloud_sql_postgres_dotnet_ado_lifetime]
-                // ConnectionIdleLifetime sets the time (in seconds) to wait before
-                // closing idle connections in the pool if the count of all
-                // connections exceeds MinPoolSize.
-                connectionString.ConnectionIdleLifetime = 300;
-                // [END cloud_sql_postgres_dotnet_ado_lifetime]
-            }
-
-            public NpgsqlConnectionStringBuilder NewPostgreSqlTCPConnectionString()
-            {
-                // [START cloud_sql_postgres_dotnet_ado_connection_tcp]
-                // Equivalent connection string:
-                // "Uid=<DB_USER>;Pwd=<DB_PASS>;Host=<DB_HOST>;Database=<DB_NAME>;"
-                var connectionString = new NpgsqlConnectionStringBuilder()
-                {
-                    // The Cloud SQL proxy provides encryption between the proxy and instance.
-                    SslMode = SslMode.Disable,
-
-                    // Remember - storing secrets in plaintext is potentially unsafe. Consider using
-                    // something like https://cloud.google.com/secret-manager/docs/overview to help keep
-                    // secrets secret.
-                    Host = Environment.GetEnvironmentVariable("DB_HOST"),     // e.g. '127.0.0.1'
-                    // Set Host to 'cloudsql' when deploying to App Engine Flexible environment
-                    Username = Environment.GetEnvironmentVariable("DB_USER"), // e.g. 'my-db-user'
-                    Password = Environment.GetEnvironmentVariable("DB_PASS"), // e.g. 'my-db-password'
-                    Database = Environment.GetEnvironmentVariable("DB_NAME"), // e.g. 'my-database'
-                };
-                connectionString.Pooling = true;
-                // Specify additional properties here.
-                // [START_EXCLUDE]
-                SetDbConfigOptions(connectionString);
-                // [END_EXCLUDE]
-                return connectionString;
-                // [END cloud_sql_postgres_dotnet_ado_connection_tcp]
-            }
-
-            NpgsqlConnectionStringBuilder NewPostgreSqlUnixSocketConnectionString()
-            {
-                // [START cloud_sql_postgres_dotnet_ado_connection_socket]
-                // Equivalent connection string:
-                // "Server=<dbSocketDir>/<INSTANCE_CONNECTION_NAME>;Uid=<DB_USER>;Pwd=<DB_PASS>;Database=<DB_NAME>"
-                String dbSocketDir = Environment.GetEnvironmentVariable("DB_SOCKET_PATH") ?? "/cloudsql";
-                String instanceConnectionName = Environment.GetEnvironmentVariable("INSTANCE_CONNECTION_NAME");
-                var connectionString = new NpgsqlConnectionStringBuilder()
-                {
-                    // The Cloud SQL proxy provides encryption between the proxy and instance.
-                    SslMode = SslMode.Disable,
-                    // Remember - storing secrets in plaintext is potentially unsafe. Consider using
-                    // something like https://cloud.google.com/secret-manager/docs/overview to help keep
-                    // secrets secret.
-                    Host = String.Format("{0}/{1}", dbSocketDir, instanceConnectionName),
-                    Username = Environment.GetEnvironmentVariable("DB_USER"), // e.g. 'my-db-user
-                    Password = Environment.GetEnvironmentVariable("DB_PASS"), // e.g. 'my-db-password'
-                    Database = Environment.GetEnvironmentVariable("DB_NAME"), // e.g. 'my-database'
-                };
-                connectionString.Pooling = true;
-                // Specify additional properties here.
-                // [START_EXCLUDE]
-                SetDbConfigOptions(connectionString);
-                // [END_EXCLUDE]
-                return connectionString;
-                // [END cloud_sql_postgres_dotnet_ado_connection_socket]
-            }
-
-            public NpgsqlConnectionStringBuilder GetPostgreSqlConnectionString()
-            {
-                if (Environment.GetEnvironmentVariable("DB_HOST") != null)
-                {
-                    return NewPostgreSqlTCPConnectionString();
-                }
-                else
-                {
-                    return NewPostgreSqlUnixSocketConnectionString();
-                }
-            }
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
